% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_predpubTable.R
\name{get.predpubTable}
\alias{get.predpubTable}
\title{Returns a dataframe with information about predatory publishers and journals from Bealls webpage}
\usage{
get.predpubTable()
}
\value{
A dataframe with publisher, name of journal and issn of predatory publications
}
\description{
This function will return a dataframe with information regarding predatory journals from Beall's site (https://scholarlyoa.com/).
The information of predatory publishers has been collected manually, while information on stand-alone journals is catpured by a web scrapping algorithm.
BE AWARE that the dataset for publishers is being built over time and it is not yet complete. Last update: 2016-12-09.
}
\section{Warning}{


While the database for standalone journals is up to date, the information about predatory publishers is not yet complete since it requires
an extensive manual work. The missing data will be regularly included in the upcoming versions of the package.
}

\examples{
my.predpub <- get.predpubTable()

head(my.predpub)
}
